+++ test/compilationTests/milestonetracker/MilestoneTracker.sol
@@ -264,10 +264,10 @@ contract MilestoneTracker {
             &&(msg.sender != recipient))
             revert();
         if (milestone.status != MilestoneStatus.AcceptedAndInProgress) revert();
-        if (now < milestone.minCompletionDate) revert();
-        if (now > milestone.maxCompletionDate) revert();
+        if (block.timestamp < milestone.minCompletionDate) revert();
+        if (block.timestamp > milestone.maxCompletionDate) revert();
         milestone.status = MilestoneStatus.Completed;
-        milestone.doneTime = now;
+        milestone.doneTime = block.timestamp;
         emit ProposalStatusChanged(_idMilestone, milestone.status);
     }
 
@@ -312,7 +312,7 @@ contract MilestoneTracker {
             &&(msg.sender != recipient))
             revert();
         if  ((milestone.status != MilestoneStatus.Completed) ||
-             (now < milestone.doneTime + milestone.reviewTime))
+             (block.timestamp < milestone.doneTime + milestone.reviewTime))
             revert();
 
         authorizePayment(_idMilestone);