+++ test/libsolidity/syntaxTests/globalFunctions/now_override.sol
@@ -1,0 +1,8 @@
+contract C {
+	function f() public view {
+		uint now = block.timestamp;
+		now;
+	}
+}
+// ----
+// Warning: (43-51): This declaration shadows a builtin symbol.